backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
Pillow==10.2.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
httpx==0.26.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.29
